* 'develop' of github.com:resource-watch/resource-watch: (45 commits)
  Add markup for second hotspot
  Add markup for hotspot 1;
  fix intro
  override cursor styles
  use pointer cursor when hovering over a hotspot
  change the position of hotspots
  Fix a bug where the "cartodb_id" field wouldn't be excluded
  Fix: tooltip interaction and wysiwyg version updated
  Style fixes
  Delete dashboards from list
  styles: minor update
  Add customized 404 page
  Upload wysiwyg version
  Add login modal that opens up when users are not logged in and they click in the button Subscribe to Alerts in Planet Pulse
  Add Login modal plus trigger from subscribe to dataset in Explore Detail
  Fix: remove console.log statement
  Update react-cesium dependency plus stop loader when the data is actually shown
  fix styles
  add second hotspot
  Remove dashboard embed
  ...

+++ components/dashboards/table/actions/DeleteAction.js
@@ -28,7 +28,7 @@ class DeleteAction extends React.Component {
 
     toastr.confirm(`Are you sure that you want to delete: "${data.name}"`, {
       onOk: () => {
-        this.service.deleteData(data.id)
+        this.service.deleteData({ id: data.id, auth: this.props.authorization })
           .then(() => {
             this.props.onRowDelete(data.id);
             toastr.success('Success', `The dashboard "${data.id}" - "${data.name}" has been removed correctly`);
